requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
pytesseract==0.3.10
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
sentencepiece==0.1.99
protobuf==4.25.1
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1.78
regex==2023.10.3